improvement(users): separate user composing logic
